UIMultilineTextInput: Placeholder should trim itself to the width of the input

+++ src/main/kotlin/club/sk1er/elementa/utils/text.kt
@@ -26,7 +26,7 @@ fun getStringSplitToWidth(text: String, maxLineWidth: Float): List<String> {
         }
 
         val word = builder.toString()
-        val wordWidth = word.width().toFloat()
+        val wordWidth = word.width()
 
         if (currLineWidth + wordWidth > maxLineWidthSpace) {
             if (wordWidth > maxLineWidthSpace) {
